Update database
Added a ternary relation from Users to Wishlishts with a perm: The table
"Accesses" with entries `uid` (a user uid), `lisht` (a wishlisht
name) and `perm` (one of "read", "write").
Added a ternary relation from Users to Users with a status: The table
"Friends" with entries `uid1` (a user uid), `uid2` (a user uid) and
`status` (one of "pending", "accepted").

Since there is no enum type in sqlite, values are curretnly not
constrained in the database.